+++ test/examples/8-localstorage-responsive.jsx
@@ -45,11 +45,11 @@ var ResponsiveLocalStorageLayout = React.createClass({
             {...this.props}
             layouts={this.state.layouts}
             onLayoutChange={this.onLayoutChange}>
-          <div key="1" _grid={{w: 2, h: 3, x: 0, y: 0}}><span className="text">1</span></div>
-          <div key="2" _grid={{w: 2, h: 3, x: 2, y: 0}}><span className="text">2</span></div>
-          <div key="3" _grid={{w: 2, h: 3, x: 4, y: 0}}><span className="text">3</span></div>
-          <div key="4" _grid={{w: 2, h: 3, x: 6, y: 0}}><span className="text">4</span></div>
-          <div key="5" _grid={{w: 2, h: 3, x: 8, y: 0}}><span className="text">5</span></div>
+          <div key="1" data-grid={{w: 2, h: 3, x: 0, y: 0}}><span className="text">1</span></div>
+          <div key="2" data-grid={{w: 2, h: 3, x: 2, y: 0}}><span className="text">2</span></div>
+          <div key="3" data-grid={{w: 2, h: 3, x: 4, y: 0}}><span className="text">3</span></div>
+          <div key="4" data-grid={{w: 2, h: 3, x: 6, y: 0}}><span className="text">4</span></div>
+          <div key="5" data-grid={{w: 2, h: 3, x: 8, y: 0}}><span className="text">5</span></div>
         </ResponsiveReactGridLayout>
       </div>
     );